Removed non-stdint types.

+++ rosserial_client/src/ros_lib/ros/duration.h
@@ -39,21 +39,21 @@
 
 namespace ros {
 
-  void normalizeSecNSecSigned(long& sec, long& nsec);
+  void normalizeSecNSecSigned(int32_t& sec, int32_t& nsec);
 
   class Duration
   {
     public:
-      long sec, nsec; 
+      int32_t sec, nsec;
       
       Duration() : sec(0), nsec(0) {}
-      Duration(long _sec, long _nsec) : sec(_sec), nsec(_nsec)
+      Duration(int32_t _sec, int32_t _nsec) : sec(_sec), nsec(_nsec)
       {
         normalizeSecNSecSigned(sec, nsec);
       }
 
       double toSec() const { return (double)sec + 1e-9*(double)nsec; };
-      void fromSec(double t) { sec = (unsigned long) floor(t); nsec = (unsigned long) round((t-sec) * 1e9); };
+      void fromSec(double t) { sec = (uint32_t) floor(t); nsec = (uint32_t) round((t-sec) * 1e9); };
 
       Duration& operator+=(const Duration &rhs);
       Duration& operator-=(const Duration &rhs);
